+++ packages/adhere-ui-conditionalrender/lib/conditionalrender.d.ts
@@ -1,5 +1,7 @@
 import PropTypes from 'prop-types';
 import { IConditionalRenderProps } from './types';
+import ConditionalRenderShow from './show';
+import ConditionalRenderVisibility from './visibility';
 /**
  * ConditionalRender - 条件渲染
  * @class ConditionalRender
@@ -7,6 +9,8 @@ import { IConditionalRenderProps } from './types';
  */
 declare function ConditionalRender({ conditional, noMatch, children }: IConditionalRenderProps): any;
 declare namespace ConditionalRender {
+    var Show: typeof ConditionalRenderShow;
+    var Visibility: typeof ConditionalRenderVisibility;
     var defaultProps: {
         conditional: boolean;
         noMatch: () => null;