+++ packages/adhere-ui-conditionalrender/changelog/CHANGELOG.md
@@ -1,7 +1,16 @@
+# 1.0.14
+
+***
+
+2022-01-25
+
+* 加入ConditionalRender.Show组件进行display切换
+* 加入ConditionalRender.Visibility组件进行visibility切换
+
 # 1.0.0
 
----
+***
 
 2020-12-22
 
-- 加入代码
+* 加入代码