+++ packages/adhere/changelog/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ***
 
-2022-1-23
+2022-1-25
 
+* **adhere-ui-conditionalrender**
+  - 加入ConditionalRender.Show组件进行display切换
+  - 加入ConditionalRender.Visibility组件进行visibility切换
 * **adhere-ui-playground**
   - 加入CodePanel组件
   - 加入CodeTabPanel组件